Columns stored as raw (memmapped) numpy arrays
This ends the experiment with column-oriented storage of data. The
conclusion is that while it's faster for loading a single column, when
loading more than a few columns (e.g., all of them), one suffers a 2x-3x
performance penalty (see the file benchmarks.txt for details)

The optimal solution seems to be to continue with row-oriented storage in
pytables tables, but separate the storage of logically connected bits (e.g.,
photometry, astrometry, etc.) into separate tablets that would be
transparently joined on fetch.