Much improved use of the logging subsystem
Each module should log to its own logger, eponymous with the module. For
example:

	logger = logging.getLogger('lsd.table')

can be found near the top of lsd/table.py

We now use YAML-based logger configuration set up in lsd.tui. Default
configuration logs warnings to console, ignores other messages. If LOG
variable is set, messages with >=INFO priority will be logged to $LOG. If
DEBUG is set, messages with >=DEBUG priority will be logged both to $LOG (if
set) and to console.

Logging can be overriden by setting LOG_CONFIG envvar to a .yml
configuration for Python's logging module.

Alternatively, the (YAML) configuration can be store to ~/.lsdlogrc.

The order in which these are looked up is as follows:

	1. LOG_CONFIG
	2. ~/.lsdlogrc
	3. Built-in default config