Support xmatching to multiple objects
Now a catalog can be x-matched to multple objects in the target catalog. For
example:

	lsd-xmatch --db=pdb --radius=60 --neighbors=10 galex_gr5 sdss

will find and record up to 10 nearest neighbors within a radius of 60 arcsec
in table sdss, around each object from table galex_gr5.

For queries to return these extra neighbors, use nmax= and dmax= arguments
to tables in FROM clause. For example:

	select  _ID, ra, dec, sdss.ra, sdss.dec
	  from  galex_gr5, sdss(nmax=20,dmax=0.5)

will return up to 20 neighbors (if >= 20 neighbors have been matched in the
first place using lsd-xmatch!) out to maximum distance of 0.5 arcsec.

If these arguments are not specified, their default values are nmax=1 and
dmax=infinity, producing the same behavior as older versions of LSD.