TabletCache: Test for root table by path, not by name
TabletCache by default loads the neighbors of all tables, unless the table
is a root table of a join. If the root table's name has been aliased to
something else (e.g., using "FROM A as B" construct), the buggy version of
the code would not recognize it as such. Now this has been fixed, as the
identity of the table is determined by path comparison, not by name.

Remaining issues:
- symlinked tables (different paths, same data) would still not compare
  equal
- if the root table is listed twice on the FROM line, it won't load
  neighbors for either occurence. This may happen for self-crossmatches,
  which are useful to find distances to nearest neighbors. Example:

	./lsd-query \
	'select ra, dec, B.ra, B.dec, _CACHED, B._NR, B._DIST*3600 \
	from ps1_obj, ps1_obj(matchedto=ps1_obj,nmax=2,dmax=30) as B'