Look up default .join files by path, not table name
Table names can be aliased (e.g. '...FROM A as B...'), leading to B being
looked up as the part of a join file name, and not A. Changed this to look
up lsd.Table.name to find the true table name.

Also, changed the code to look up default .join files only in the dbdir
containing the table in question, and not in the entire LSD_DB path.