Override "back_to_disk" with LSD_DISKLESS environment variable
If LSD_DISKLESS environment variable is set to 1, the "back_to_disk" variable in pool2 will be set to False.
Otherwsie, "back_to_disk" is set to True (default), unless the filesytem is HFS+.
This feature allows the user to run LSD on resources that do not have disks or that do not have filesystems
that support memory mapping on disks (such as GPFS).

Signed-off-by: Mario Juric <mjuric@cfa.harvard.edu>