Implement atomic, consistent, isolated and durable transactions
All changes (writes) to LSD tables are now performed in snapshot-based ACID
transactions.  Until a transaction commits, the modifications that were made
in the transaction will not be visible to readers. On commit, the neighbor
caches and tablet tree caches (soon to be renamed to 'catalogs') are
automatically (and intelligently) rebuilt, ensuring consistency.

LSD implements transactions using a variant of the snapshot isolation
technique. Each LSD table has a 'snapshots' directory, with subdirectories
storing snapshot data. Snapshots can either be opened or committed; a
committed snapshot contains special file '.committed', as a marker of its
state.

The data logically contained in the table consists of a union of contents of
all committed snapshot directories, made from oldest to newest committed
snapshot, where contents (files) of newer snapshots overwrite eponymous
files from older ones.

As a consequence of this implementation:

    * Rolling back to an older snapshot can be achieved by removing
      directories containing newer snapshots. Actually, in principle the
      directories don't even have to be removed -- LSD just needs to be told
      to look for a specific snapshot -- but this is not implemented yet.

    * To read a given snapshots, all older snapshots must be present. You
      can view each snapshot as a "diff" between the current and previous
      state of the database, going back to the beginning; all diffs have to
      be present to construct the current state.

    * If anything goes wrong in a transaction, the snapshot directory
      created by the transaction will be left in the snapshots/
      subdirectory, but won't have a '.committed' file, and therefore be
      ignored by LSD. They can be safely removed, either manually ('rm
      -rf'), or using lsd-vacuum (still to be written).

    * Queries to the database don't see the data added by the current
      transaction; they see the database state as it was when the
      transaction was started. For example, if you have a table with 10
      rows, begin a transaction, add or modify some rows and, without
      committing, query that table again, you will get the original 10 rows
      as a result. Only after you've called db.commit() will your queries
      begin returning the new data.

    * Upon commit, LSD will do the necessary housekeeping, including the
      updating of table catalogs (catalog.pkl), as well as intelligently
      updating the neighbor caches for the cells that were modified by the
      transaction.