Added lsd.locking module, with file-based locks
This should be a replacement for procmail'd lockfile. I've tested it to work
over NFS and Lustre (see the test files in attic/locking directory).

The context manager locking.lock guarantees that no lock will remain
acquired if an exception is thrown anywhere in the with statement (or in the
code attempting to acquire the lock).

However, this will not be the case if a signal is caught while acquiring the
lock (e.g.  hitting CTRL-C on the console).  Lockfiles may remain after the
program has exited due to the signal, and will have to be removed manually.
Ensuring this does not happen, and that lockfiles are always cleaned up even
in face of signals, is virtually impossible in pure Python because of lack
of signal blocking.