lsd.import: fix imports with defined primary key
Also extend the make_image_cache() commit hook with an argument to specify
whether all detection table cells should be inspected or just those from the
latest snapshot.  This is a (very dirty) hack for Branimir Sesar, to permit
him to import a table of zero-points that is qualitatively handled as
another table of exposures.