Introduce utils.NamedList, add equgal()
Disallow in-query callables to return a list of (name, column) tuples,
instead have them use NamedList. Given how NamedList class is written, this
means that when iterated over it behaves exactly as if names were not there,
thus allowing things like '... equgal(*galequ(l, b)) ...' to be written in
queries.