Added lsd-data as a git submodule in src/data
lsd-data is kept separately from the main github repository, because of
the size of the files it holds (github quota is too small to accommodate
them). They're on:

	http://faun.rc.fas.harvard.edu/mjuric/lsd-data.git

Note that git submodules are not checked out (downloaded) by default. To get
the contents of src/data do:

	git submodule init
	git submodule update

This will download the data files from the URL above.